+++ lib/weighted_random/probs/probs.ex
@@ -1,0 +1,3 @@
+defmodule WeightedRandom.Probs do
+  @moduledoc false
+end