+++ test/weighted_random/backends/linear_test.exs
@@ -3,8 +3,7 @@ defmodule WeightedRandom.Backends.LinearTest do
   alias WeightedRandom.Backend.Linear, as: Mod
 
   test "Index true" do
-    opts = [backend: Mod]
-           |> WeightedRandom.Utils.Opts.merge_opts()
+    opts = WeightedRandom.Input.Opts.from_weights_merge_opts([backend: Mod])
 
     input = WeightedRandom.Input.FromWeights.get_inputs(200..300, [%{target: 5, weight: 25}], opts)
     assert is_struct(input, WeightedRandom.Input)
@@ -19,18 +18,20 @@ defmodule WeightedRandom.Backends.LinearTest do
   end
 
   test "index false" do
+    outcomes = 200..300
     target = 205
+    outcome_type = :value
     weight = 25
 
-    opts = [backend: Mod, index: false]
-           |> WeightedRandom.Utils.Opts.merge_opts()
+    opts = [backend: Mod, outcome_type: outcome_type]
+           |> WeightedRandom.Input.Opts.from_weights_merge_opts()
 
-    input = WeightedRandom.Input.FromWeights.get_inputs(200..300, [%{target: target, weight: weight}], opts)
+    input = WeightedRandom.Input.FromWeights.get_inputs(outcomes, [%{target: target, weight: weight}], opts)
     assert is_struct(input, WeightedRandom.Input)
 
     table = Mod.preprocess(input, opts)
     assert is_struct(table, Mod)
-    assert Enum.count(table.li) == Enum.count(200..300) + weight
+    assert Enum.count(table.li) == Enum.count(outcomes) + weight
 
     li = Mod.take(table, 400)
     assert Enum.all?(li, &(&1 >= 200))